[HO] Code health.
- Breaks down instance manager into smaller independent units catching
  up with the Single Responsability Principle.
