+++ frontend/src/host_orchestrator/orchestrator/listcvdsaction.go
@@ -1,0 +1,58 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package orchestrator
+
+import (
+	"github.com/google/android-cuttlefish/frontend/src/host_orchestrator/orchestrator/cvd"
+
+	apiv1 "github.com/google/android-cuttlefish/frontend/src/liboperator/api/v1"
+)
+
+type ListCVDsActionOpts struct {
+	Paths           IMPaths
+	ExecContext     ExecContext
+	CVDToolsVersion AndroidBuild
+	BuildAPIFactory BuildAPIFactory
+	CVDUser         string
+}
+
+type ListCVDsAction struct {
+	paths              IMPaths
+	downloadCVDHandler *downloadCVDHandler
+	execContext        cvd.CVDExecContext
+}
+
+func NewListCVDsAction(opts ListCVDsActionOpts) *ListCVDsAction {
+	return &ListCVDsAction{
+		paths: opts.Paths,
+		downloadCVDHandler: newDownloadCVDHandler(
+			opts.CVDToolsVersion,
+			&opts.Paths,
+			opts.BuildAPIFactory(""), // cvd can be downloaded without credentials
+		),
+		execContext: newCVDExecContext(opts.ExecContext, opts.CVDUser),
+	}
+}
+
+func (a *ListCVDsAction) Run() (*apiv1.ListCVDsResponse, error) {
+	if err := a.downloadCVDHandler.Download(); err != nil {
+		return nil, err
+	}
+	fleet, err := cvdFleet(a.execContext, a.paths.CVDBin())
+	if err != nil {
+		return nil, err
+	}
+	return &apiv1.ListCVDsResponse{CVDs: fleetToCVDs(fleet)}, nil
+}
