+++ frontend/src/host_orchestrator/orchestrator/listcvdsaction_test.go
@@ -1,0 +1,87 @@
+// Copyright 2023 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package orchestrator
+
+import (
+	"context"
+	"os/exec"
+	"path"
+	"strings"
+	"testing"
+
+	orchtesting "github.com/google/android-cuttlefish/frontend/src/host_orchestrator/orchestrator/testing"
+	apiv1 "github.com/google/android-cuttlefish/frontend/src/liboperator/api/v1"
+
+	"github.com/google/go-cmp/cmp"
+)
+
+func TestListCVDsSucceeds(t *testing.T) {
+	dir := orchtesting.TempDir(t)
+	defer orchtesting.RemoveDir(t, dir)
+	output := `[
+  [
+          {
+                  "adb_serial" : "0.0.0.0:6520",
+                  "assembly_dir" : "/var/lib/cuttlefish-common/runtimes/cvd-1/cuttlefish/assembly",
+                  "displays" :
+                  [
+                          "720 x 1280 ( 320 )"
+                  ],
+                  "instance_dir" : "/var/lib/cuttlefish-common/runtimes/cvd-1/cuttlefish/instances/cvd-1",
+                  "instance_name" : "cvd-1",
+                  "status" : "Running",
+                  "web_access" : "https:///run/cuttlefish/operator:8443/client.html?deviceId=cvd-1",
+                  "webrtc_port" : "8443"
+          }
+  ]
+]`
+	execContext := func(ctx context.Context, name string, args ...string) *exec.Cmd {
+		cmd := exec.Command("true")
+		if path.Base(args[len(args)-1]) == "fleet" {
+			// Prints to stderr as well to make sure only stdout is used.
+			cmd = exec.Command("tee", "/dev/stderr")
+			cmd.Stdin = strings.NewReader(strings.TrimSpace(output))
+		}
+		return cmd
+	}
+	cvdBinAB := AndroidBuild{ID: "1", Target: "xyzzy"}
+	paths := IMPaths{
+		CVDToolsDir:      dir,
+		ArtifactsRootDir: dir + "/artifacts",
+		RuntimesRootDir:  dir + "/runtimes",
+	}
+	opts := ListCVDsActionOpts{
+		Paths:           paths,
+		ExecContext:     execContext,
+		CVDToolsVersion: cvdBinAB,
+		BuildAPIFactory: func(_ string) BuildAPI { return &fakeBuildAPI{} },
+		CVDUser:         fakeCVDUser,
+	}
+	action := NewListCVDsAction(opts)
+
+	res, _ := action.Run()
+
+	want := &apiv1.ListCVDsResponse{CVDs: []*apiv1.CVD{
+		{
+			Name:        "cvd-1",
+			BuildSource: &apiv1.BuildSource{},
+			Status:      "Running",
+			Displays:    []string{"720 x 1280 ( 320 )"},
+		},
+	}}
+	if diff := cmp.Diff(want, res); diff != "" {
+		t.Errorf("response mismatch (-want +got):\n%s", diff)
+	}
+}
